+++ opsml/projects/_run_manager.py
@@ -8,7 +8,7 @@ from typing import Dict, Optional, Union, cast
 
 from opsml.helpers.logging import ArtifactLogger
 from opsml.projects._active_run import ActiveRun, RunInfo
-from opsml.projects.types import ProjectInfo, Tags
+from opsml.projects.base.types import ProjectInfo, Tags
 from opsml.registry import CardRegistries, CardRegistry, ProjectCard, RunCard
 from opsml.registry.storage.storage_system import StorageClientType
 from opsml.registry.utils.settings import settings
